add separate file for product types
