+++ app/types/product.ts
@@ -1,0 +1,8 @@
+export type APIProducts = {
+  id: number
+  name: string
+  description: string
+  price: number
+  image: string
+  rating: number
+}
